proxy/requirements.txt
mitmproxy==9.0.1
requests==2.31.0
python-dotenv==1.0.0